blkdevice: main_time is 64-bit — int param wrapped at 2^31 and killed the disk
BeforeEval(int cycles) received the 64-bit half-cycle counter; at
exactly 2,147,483,648 half-cycles (~1.07G machine cycles) it went
negative and the 'if (cycles<2000) return' boot guard disabled the
block device permanently — mid-results-write on the gate run: io_wr
held forever, no io_ack, ROM in the eternal INT poll at $40899704.
Found via the new NCR command/INT/io-edge $display taps (kept,
VERILATOR-gated): everything healthy for 2M byte-writes, then one
flush with no ack at the 2^31 boundary.

Same genus as finding 30 ('Verilog unsized literals are 32-bit'),
C++ edition.  Any long sim run with disk I/O past ~1.07G machine
cycles hit this — including, likely, the desktop's corpus reruns.

+++ verilator/sim/sim_blkdevice.cpp
@@ -76,7 +76,7 @@ void SimBlockDevice::MountDisk( std::string file, int index) {
 }
 
 
-void SimBlockDevice::BeforeEval(int cycles)
+void SimBlockDevice::BeforeEval(long long cycles)
 {
 //
 // switch to a new disk if current_disk is -1
